requirement_dev.txt
arize-phoenix==8.26.0
chromadb==1.0.6
evidently==0.7.0
greenlet==3.2.0
httptools==0.6.4
langchain-groq==0.3.2
langchain-mistralai==0.2.10
openinference-instrumentation-langchain==0.1.41
pip==25.0.1
ragas==0.2.14
rouge_score==0.1.2
sacrebleu==2.5.1
sentence-transformers==4.1.0
uvloop==0.21.0
watchfiles==1.0.5
